Fix attachments not serializing properly (take 4)
1. Split MessageAttachment and MessageAttachmentBody
2. Prettify whitespace in generated types
3. Minor refactoring in generator

+++ vk-api-methods/src/main/kotlin/name/anton3/vkapi/methods/longpoll/AbstractLongPollEventSource.kt
@@ -45,11 +45,11 @@ abstract class AbstractLongPollEventSource<EventType, IteratorType>(
 
         return when (lpResponse.failed) {
             null -> {
-                iteratorWithTs(iterator, lpResponse.ts!!) to lpResponse.updates
+                iteratorWithTs(iterator, lpResponse.ts!!.toInt()) to lpResponse.updates
             }
             LongPollFailure.NEW_TS -> {
                 log.debug("LongPoll failed with error 1")
-                iteratorWithTs(iterator, lpResponse.ts!!) to emptyList()
+                iteratorWithTs(iterator, lpResponse.ts!!.toInt()) to emptyList()
             }
             LongPollFailure.REQUEST_NEW_KEY -> {
                 val newServer = iterator()
